agents/requirements.txt
psutil>=5.9.0
py-cpuinfo>=9.0.0
netifaces>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
pycrypto>=2.6.1
compression>=0.1.5
tenacity>=8.2.3
pytest>=7.4.0
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0
